% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followButton.R
\name{followButton}
\alias{followButton}
\title{Add follow Button to any page}
\usage{
followButton(
  link,
  image = NULL,
  text = NULL,
  position = c("left", "right", "inline"),
  text.color = "black"
)
}
\arguments{
\item{link}{the link to share on social media}

\item{image}{the image link to share on social media}

\item{text}{the text to share on social media}

\item{position}{of buttons e.g "left","right","bottom","inline"}

\item{text.color}{text color e.g black}
}
\value{
Add follow button to any page
}
\description{
Customize and include follow button in markdown and shiny apps
}
\section{Examples for r2social}{

More examples and demo pages are located at this link -
\url{https://api.rpkg.net}.
}

\examples{

followButton(link = "http://rpkg.net", position = "left")
followButton(link = "http://obianom.com", position = "right")
followButton(link = "http://66pharm.com", position = "inline")

}
