% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likeButton.R
\name{likeButton}
\alias{likeButton}
\title{Add Like Button to any page}
\usage{
likeButton(text = "Like", text.color = "black", count = TRUE)
}
\arguments{
\item{text}{the text to share on social media}

\item{text.color}{text color e.g black}

\item{count}{count of number of likes}
}
\value{
Add like button to any page
}
\description{
Customize and include like button in markdown and shiny apps
}
\section{Examples for r2social}{

More examples and demo pages are located at this link -
\url{https://api.rpkg.net}.
}

\examples{

likeButton()

}
