% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubpprofile.R
\name{profileDisplay1}
\alias{profileDisplay1}
\title{Display Github Profile}
\usage{
profileDisplay1(item)
}
\arguments{
\item{item}{list of profiles}
}
\description{
Social card for a github profile
}
\note{
Obtain GitHub API at https://github.com/settings/tokens, make sure to login
}
