% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialtools.R
\name{r2social.scripts}
\alias{r2social.scripts}
\title{Add r2social scripts}
\usage{
r2social.scripts(name = "sharesocial")
}
\arguments{
\item{name}{script file name}
}
\value{
scripts needed for styling
}
\description{
Wrap functions needed for styling of the containers
}
\examples{

if (interactive()) {
  r2social.scripts()
}

}
