% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratingButton.R
\name{ratingButton}
\alias{ratingButton}
\title{Add rating Button to any page}
\usage{
ratingButton(text = "rating", text.color = "black", count = TRUE)
}
\arguments{
\item{text}{the text to share on social media}

\item{text.color}{text color e.g black}

\item{count}{count of number of ratings}
}
\value{
Add rating button to any page
}
\description{
Customize and include rating button in markdown and shiny apps
}
\section{Examples for r2social}{

More examples and demo pages are located at this link -
\url{https://api.rpkg.net}.
}

\examples{

ratingButton()

}
