% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialtools.R
\name{shareButton}
\alias{shareButton}
\title{Add social buttons to share a page}
\usage{
shareButton(
  link,
  image = NULL,
  text = NULL,
  position = c("left", "right", "bottom", "inline"),
  text.color = "black",
  facebook = TRUE,
  linkedin = TRUE,
  twitter = TRUE,
  tumblr = TRUE,
  pinterest = TRUE,
  whatsapp = TRUE,
  reddit = TRUE,
  instagram = TRUE,
  blogger = TRUE,
  weibo = TRUE,
  tiktok = TRUE,
  vk = TRUE,
  telegram = TRUE,
  youtube = TRUE
)
}
\arguments{
\item{image}{= NULL,}

\item{text}{= NULL,}

\item{position}{= c("left","right","bottom","inline"),}

\item{text.color}{= "black",}

\item{facebook}{share on Facebook}

\item{linkedin}{share on Linkedin}

\item{twitter}{share on Twitter}

\item{tumblr}{share on Tumblr}

\item{pinterest}{share on Pinterest}

\item{whatsapp}{share on Whatsapp}

\item{reddit}{share on Reddit}

\item{instagram}{share on Instagram}

\item{blogger}{share on blogger}

\item{weibo}{share on weibo}

\item{tiktok}{share on tiktok}

\item{vk}{share on VK or VKontakte}

\item{telegram}{share on Telegram}

\item{youtube}{share on Youtube}
}
\value{
Resizeable split screen container
}
\description{
Customize social buttons available for sharing of pages
}
\section{Examples for r2social}{

More examples and demo pages are located at this link -
\url{https://rpkg.net/package/r2social}.
}

\examples{

shareButton()

}
