% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialtools.R
\name{sharebuttons}
\alias{sharebuttons}
\title{Add social buttons to a page}
\usage{
sharebuttons(
  ...,
  position = c("vertical", "horizontal"),
  bg.color = "black",
  text.color = "black",
  muisc = NULL
)
}
\arguments{
\item{...}{list of social buttons to add}

\item{position}{position of the share buttons}

\item{bg.color}{background color}

\item{text.color}{text color}

\item{misc}{miscillanenous}
}
\value{
Resizeable split screen container
}
\description{
Highly customizable social buttons for the pages
}
\section{Examples for r2social}{

More examples and demo pages are located at this link -
\url{https://rpkg.net/package/r2social}.
}

\examples{

sharebuttons()

}
