% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialtools.R
\name{socialButtons}
\alias{socialButtons}
\title{Customizable social icons}
\usage{
socialButtons(
  link,
  type = c("share", "connect"),
  image = NULL,
  text = NULL,
  position = c("left", "right", "inline"),
  text.color = "white",
  facebook = FALSE,
  linkedin = FALSE,
  twitter = FALSE,
  tumblr = FALSE,
  pinterest = FALSE,
  whatsapp = FALSE,
  reddit = FALSE,
  instagram = FALSE,
  blogger = FALSE,
  weibo = FALSE,
  tiktok = FALSE,
  vk = FALSE,
  telegram = FALSE,
  youtube = FALSE
)
}
\arguments{
\item{link}{link address to share}

\item{type}{type of social button e.g. share or connect}

\item{image}{image link for pinterest only}

\item{text}{text link for twitter only}

\item{position}{= c("left","right","bottom","inline"),}

\item{text.color}{= "black",}

\item{facebook}{share on Facebook}

\item{linkedin}{share on Linkedin}

\item{twitter}{share on Twitter}

\item{tumblr}{share on Tumblr}

\item{pinterest}{share on Pinterest}

\item{whatsapp}{share on Whatsapp}

\item{reddit}{share on Reddit}

\item{instagram}{share on Instagram}

\item{blogger}{share on blogger}

\item{weibo}{share on weibo}

\item{tiktok}{share on tiktok}

\item{vk}{share on VK or VKontakte}

\item{telegram}{share on Telegram}

\item{youtube}{share on Youtube}
}
\value{
Resizeable split screen container
}
\description{
For use in making buttons to share a page or connect to your social media
}
\section{Examples for r2social}{

More examples and demo pages are located at this link -
\url{https://rpkg.net/package/r2social}.
}

\examples{

socialButtons("https://66pharm.com")

}
