% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2social-package.R
\docType{package}
\name{r2social-package}
\alias{r2social}
\alias{r2social-package}
\title{r2social: Seamless Integration of Sharing and Connect Buttons in Markdown and Apps}
\description{
Implementation of 'JQuery' \url{https://jquery.com} and 'CSS' styles to allow easy incorporation of various social media elements on a page. The elements include addition of share buttons or connect with us buttons or hyperlink buttons to 'Shiny' applications or dashboards and 'Rmarkdown' documents.Sharing capability on social media platforms including 'Facebook' \url{https://www.facebook.com}, 'Linkedin' \url{https://www.linkedin.com}, 'X/Twitter' \url{https://x.com}, 'Tumblr' \url{https://www.tumblr.com}, 'Pinterest' \url{https://www.pinterest.com}, 'Whatsapp' \url{https://www.whatsapp.com}, 'Reddit' \url{https://www.reddit.com}, 'Baidu' \url{https://www.baidu.com}, 'Blogger' \url{https://www.blogger.com}, 'Weibo' \url{https://www.weibo.com}, 'Instagram' \url{https://www.instagram.com}, 'Telegram' \url{https://www.telegram.me}, 'Youtube' \url{https://www.youtube.com}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r2social.obi.obianom.com}
  \item Report bugs at \url{https://github.com/oobianom/r2social/issues}
}

}
\author{
\strong{Maintainer}: Obinna Obianom \email{idonshayo@gmail.com}

}
\keyword{internal}
