% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2social-package.R
\docType{package}
\name{r2social-package}
\alias{r2social}
\alias{r2social-package}
\title{r2social: Social Buttons Inclusion on a Page}
\description{
Add social buttons on 'Shiny', 'Markdown', 'Rmarkdown' and 'Quarto' documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r2social.obi.obianom.com/}
  \item Report bugs at \url{https://github.com/oobianom/r2social/issues}
}

}
\keyword{internal}
